rename test files, merge pytest_utils with utils
