+++ tests/python/opinfo_core.py
