+++ tests/python/opinfo_fusion_definitions.py
