+++ tests/python/opinfo_input_generators.py
