+++ tests/python/opinfos.py
